after_review

+++ recipes/views.py
@@ -1,8 +1,8 @@
 import json
 from urllib.parse import unquote
 
-from django.conf import settings
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import ObjectDoesNotExist
 from django.core.paginator import Paginator
 from django.db.models import Sum
 from django.http import HttpResponse, JsonResponse
@@ -11,18 +11,30 @@ from django.utils.decorators import method_decorator
 from django.views import View
 from django.views.decorators.http import (require_GET, require_http_methods,
                                           require_POST)
-
 from users.models import Subscription
 
 from .forms import RecipeForm
-from .managers import (Favorite, Purchase, add_subscription_status,
-                       extend_context)
-from .models import Ingredient, Product, Recipe, Tag, User
+from .models import Favorite, Ingredient, Product, Purchase, Recipe, Tag, User
+
+
+def extend_context(context, user):
+    context['purchase_list'] = Purchase.purchase.get_purchases_list(user)
+    context['favorites'] = Favorite.favorite.get_favorites(user)
+    return context
+
+
+def add_subscription_status(context, user, author):
+    context['is_subscribed'] = Subscription.objects.filter(
+        user=user, author=author
+    ).exists()
+    return context
 
 
 @require_GET
 def index(request):
-    paginator = Paginator(settings.PAGINATION_PAGE_SIZE)
+    tags = request.GET.getlist('tag')
+    recipe_list = Recipe.recipes.tag_filter(tags)
+    paginator = Paginator(recipe_list, 6)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
     context = {
@@ -42,10 +54,7 @@ def profile(request, user_id):
     profile = get_object_or_404(User, id=user_id)
     tags = request.GET.getlist('tag')
     recipes_list = Recipe.recipes.tag_filter(tags)
-    paginator = Paginator(
-        recipes_list.filter(author=profile),
-        settings.PAGINATION_PAGE_SIZE
-    )
+    paginator = Paginator(recipes_list.filter(author=profile), 6)
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
     context = {
@@ -60,7 +69,7 @@ def profile(request, user_id):
     if user.is_authenticated:
         add_subscription_status(context, user, profile)
         extend_context(context, user)
-    return render(request, 'profile.html', context)
+    return render(request, 'recipes/profile.html', context)
 
 
 @require_GET
@@ -73,7 +82,7 @@ def recipe_detail(request, recipe_id):
     if user.is_authenticated:
         add_subscription_status(context, user, recipe.author)
         extend_context(context, user)
-    return render(request, 'recipe_detail.html', context)
+    return render(request, 'recipes/recipe_detail.html', context)
 
 
 class FavoriteView(View):
@@ -90,10 +99,7 @@ class FavoriteView(View):
         return queryset
 
     def get(self, request):
-        paginator = Paginator(
-            self.get_queryset(),
-            settings.PAGINATION_PAGE_SIZE
-        )
+        paginator = Paginator(self.get_queryset(), 6)
         page_number = request.GET.get('page')
         page = paginator.get_page(page_number)
         purchase_list = Purchase.purchase.get_purchases_list(request.user)
@@ -104,22 +110,17 @@ class FavoriteView(View):
             'paginator': paginator,
             'page': page
         }
-        return render(request, 'favorites.html', context)
+        return render(request, 'recipes/favorites.html', context)
 
     def post(self, request):
         json_data = json.loads(request.body.decode())
-        recipe_id = []
-        if json_data['id'] is True:
-            try:
-                recipe_id = json_data['id']
-            except KeyError:
-                recipe_id = None
+        recipe_id = json_data['id']
         recipe = get_object_or_404(Recipe, id=recipe_id)
-        data = {'success': True}
+        data = {'success': 'true'}
         favorite = Favorite.favorite.get_user(request.user)
         is_favorite = favorite.recipes.filter(id=recipe_id).exists()
         if is_favorite:
-            data['success'] = False
+            data['success'] = 'false'
         else:
             favorite.recipes.add(recipe)
         return JsonResponse(data)
@@ -129,13 +130,13 @@ class FavoriteView(View):
 @require_http_methods('DELETE')
 def delete_favorite(request, recipe_id):
     recipe = get_object_or_404(Recipe, id=recipe_id)
-    data = {'success': True}
+    data = {'success': 'true'}
     try:
         favorite = Favorite.favorite.get(user=request.user)
-    except User.DoesNotExist:
-        data['success'] = False
+    except ObjectDoesNotExist:
+        data['success'] = 'false'
     if not favorite.recipes.filter(id=recipe_id).exists():
-        data['success'] = False
+        data['success'] = 'false'
     favorite.recipes.remove(recipe)
     return JsonResponse(data)
 
@@ -147,19 +148,17 @@ def get_subscriptions(request):
         subscriptions = Subscription.objects.filter(
             user=request.user
         ).order_by('pk')
-    except User.DoesNotExist:
+    except ObjectDoesNotExist:
         subscriptions = []
     page_num = request.GET.get('page')
-    paginator = Paginator(
-        subscriptions,
-        settings.PAGINATION_PAGE_SIZE)
+    paginator = Paginator(subscriptions, 6)
     page = paginator.get_page(page_num)
     context = {
         'active': 'subscription',
         'paginator': paginator,
         'page': page,
     }
-    return render(request, 'subscriptions.html', context)
+    return render(request, 'recipes/subscriptions.html', context)
 
 
 @login_required(login_url='auth/login/')
@@ -169,9 +168,9 @@ def subscription(request):
     author = get_object_or_404(User, id=json_data['id'])
     is_exist = Subscription.objects.filter(
         user=request.user, author=author).exists()
-    data = {'success': True}
+    data = {'success': 'true'}
     if is_exist:
-        data['success'] = False
+        data['success'] = 'false'
     else:
         Subscription.objects.create(user=request.user, author=author)
     return JsonResponse(data)
@@ -181,11 +180,11 @@ def subscription(request):
 @require_http_methods('DELETE')
 def delete_subscription(request, author_id):
     author = get_object_or_404(User, id=author_id)
-    data = {'success': True}
+    data = {'success': 'true'}
     follow = Subscription.objects.filter(
         user=request.user, author=author)
     if not follow:
-        data['success'] = False
+        data['success'] = 'false'
     follow.delete()
     return JsonResponse(data)
 
@@ -207,7 +206,7 @@ class PurchaseView(View):
             'recipes_list': recipes_list,
             'active': 'purchase'
         }
-        return render(request, 'purchases.html', context)
+        return render(request, 'recipes/purchases.html', context)
 
     def post(self, request):
         json_data = json.loads(request.body.decode())
@@ -215,12 +214,12 @@ class PurchaseView(View):
         recipe = get_object_or_404(Recipe, id=recipe_id)
         purchase = Purchase.purchase.get_user_purchase(user=request.user)
         data = {
-            'success': True
+            'success': 'true'
         }
         if not purchase.recipes.filter(id=recipe_id).exists():
             purchase.recipes.add(recipe)
             return JsonResponse(data)
-        data['success'] = False
+        data['success'] = 'false'
         return JsonResponse(data)
 
 
@@ -229,14 +228,14 @@ class PurchaseView(View):
 def delete_purchase(request, recipe_id):
     recipe = get_object_or_404(Recipe, id=recipe_id)
     data = {
-        'success': True
+        'success': 'true'
     }
     try:
         purchase = Purchase.purchase.get(user=request.user)
-    except User.DoesNotExist:
-        data['success'] = False
+    except ObjectDoesNotExist:
+        data['success'] = 'false'
     if not purchase.recipes.filter(id=recipe_id).exists():
-        data['success'] = False
+        data['success'] = 'false'
     purchase.recipes.remove(recipe)
     return JsonResponse(data)
 
@@ -264,33 +263,40 @@ def send_shop_list(request):
 
 
 @login_required(login_url='auth/login/')
+@require_http_methods(['GET', 'POST'])
 def new_recipe(request):
     context = {
         'active': 'new_recipe',
         'page_title': 'Создание рецепта',
         'button_label': 'Создать рецепт',
     }
-    form = RecipeForm(request.POST or None, files=request.FILES or None)
-    if not form.is_valid():
+    # GET-запрос на страницу создания рецепта
+    if request.method == 'GET':
+        form = RecipeForm()
         context['form'] = form
-        return render(request, 'recipe_form.html', context)
-    recipe = form.save(commit=False)
-    recipe.author = request.user
-    form.save()
-    ingedient_names = request.POST.getlist('nameIngredient')
-    ingredient_units = request.POST.getlist('unitsIngredient')
-    amounts = request.POST.getlist('valueIngredient')
-    products = [get_object_or_404(
-        Product.objects.all,
-        title=ingedient_names[i],
-        unit=ingredient_units[i]
-    ) for i in range(len(ingedient_names))]
-    ingredients = []
-    for i in zip(recipe, products, amounts):
-        ingredients.append(Ingredient(
-            recipe=recipe, ingredient=products[i], amount=amounts[i]))
-    Ingredient.objects.bulk_create(ingredients)
-    return redirect('index')
+        return render(request, 'recipes/recipe_form.html', context)
+    # POST-запрос с данными из формы создания рецепта
+    elif request.method == 'POST':
+        form = RecipeForm(request.POST, files=request.FILES or None)
+        if not form.is_valid():
+            context['form'] = form
+            return render(request, 'recipes/recipe_form.html', context)
+        recipe = form.save(commit=False)
+        recipe.author = request.user
+        form.save()
+        ingedient_names = request.POST.getlist('nameIngredient')
+        ingredient_units = request.POST.getlist('unitsIngredient')
+        amounts = request.POST.getlist('valueIngredient')
+        products = [Product.objects.get(
+            title=ingedient_names[i],
+            unit=ingredient_units[i]
+        ) for i in range(len(ingedient_names))]
+        ingredients = []
+        for i in range(len(amounts)):
+            ingredients.append(Ingredient(
+                recipe=recipe, ingredient=products[i], amount=amounts[i]))
+        Ingredient.objects.bulk_create(ingredients)
+        return redirect('index')
 
 
 @login_required(login_url='auth/login/')
@@ -305,6 +311,7 @@ def get_ingredients(request):
 
 
 @login_required(login_url='auth/login/')
+@require_http_methods(['GET', 'POST'])
 def edit_recipe(request, recipe_id):
     recipe = get_object_or_404(Recipe, id=recipe_id)
     context = {
@@ -312,36 +319,40 @@ def edit_recipe(request, recipe_id):
         'page_title': 'Редактирование рецепта',
         'button_label': 'Сохранить',
     }
-    form = RecipeForm(request.POST or None, files=request.FILES or None)
-    if not form.is_valid():
+    # GET-запрос на страницу редактирования рецепта
+    if request.method == 'GET':
+        form = RecipeForm(instance=recipe)
         context['form'] = form
-        return render(request, 'recipe_form.html', context)
-    recipe = form.save(commit=False)
-    recipe.author = request.user
-    form.save()
-    new_titles = request.POST.getlist('nameIngredient')
-    new_units = request.POST.getlist('unitsIngredient')
-    amounts = request.POST.getlist('valueIngredient')
-    Ingredient.objects.filter(recipe__id=recipe_id).delete()
-    product = [get_object_or_404(
-        Product.objects.all,
-        title=new_titles[i],
-        unit=new_units[i]
-    ) for i in range(len(new_titles))]
-    new_ingredients = []
-    for i in zip(recipe, product, amounts):
-        new_ingredients.append(Ingredient(
-            recipe=recipe, ingredient=product[i], amount=amounts[i]))
-    Ingredient.objects.bulk_create(new_ingredients)
-    return redirect('index')
+        return render(request, 'recipes/recipe_form.html', context)
+    # POST-запрос с данными из формы редактирования рецепта
+    elif request.method == 'POST':
+        form = RecipeForm(request.POST or None,
+                          files=request.FILES or None, instance=recipe)
+        if not form.is_valid():
+            context['form'] = form
+            return render(request, 'recipes/recipe_form.html', context)
+        form.save()
+        new_titles = request.POST.getlist('nameIngredient')
+        new_units = request.POST.getlist('unitsIngredient')
+        amounts = request.POST.getlist('valueIngredient')
+        products_num = len(new_titles)
+        new_ingredients = []
+        Ingredient.objects.filter(recipe__id=recipe_id).delete()
+        for i in range(products_num):
+            product = Product.objects.get(
+                title=new_titles[i], unit=new_units[i])
+            new_ingredients.append(Ingredient(recipe=recipe,
+                                              ingredient=product,
+                                              amount=amounts[i]))
+        Ingredient.objects.bulk_create(new_ingredients)
+        return redirect('index')
 
 
 @login_required(login_url='auth/login/')
 @require_GET
 def delete_recipe(request, recipe_id):
     recipe = get_object_or_404(Recipe, id=recipe_id)
-    if request.user.is_superuser or request.user == recipe.author:
-        recipe.delete()
+    recipe.delete()
     return redirect('index')
